% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallrFutureBackend-class.R
\name{CallrFutureBackend}
\alias{CallrFutureBackend}
\title{A callr future is a future whose value will be resolved via callr}
\usage{
CallrFutureBackend(workers = availableCores(), supervise = FALSE, ...)
}
\arguments{
\item{workers}{(optional) The maximum number of workers the callr
backend may use at any time.}

\item{supervise}{(optional) Argument passed to \code{\link[callr:r_bg]{callr::r_bg()}}.}

\item{\ldots}{Additional arguments passed to \code{\link[future:FutureBackend]{future::FutureBackend()}}.}
}
\value{
A CallrFutureBackend object
}
\description{
A callr future is a future whose value will be resolved via callr
}
\keyword{internal}
