% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallrFutureError-class.R
\name{CallrFutureError}
\alias{CallrFutureError}
\title{FutureError class for errors related to CallrFuture:s}
\usage{
CallrFutureError(...)
}
\arguments{
\item{\ldots}{Arguments passed to \link[future:FutureCondition]{FutureError}.}
}
\description{
FutureError class for errors related to CallrFuture:s
}
\keyword{internal}
