% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{future.processx}
\alias{future.processx}
\alias{future.processx-package}
\title{future.processx: A Future for processx}
\description{
The \pkg{future.processx} package implements the Future API
on top of \pkg{processx}.
The Future API is defined by the \pkg{future} package.
}
\details{
To use processx futures, load \pkg{future.processx}, and
select the type of future you wish to use, e.g. \code{plan(processx)}.
}
\examples{
\donttest{
plan(processx)
demo("mandelbrot", package = "future", ask = FALSE)
}

}
