% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallrFutureBackend-class.R
\name{print.CallrFuture}
\alias{print.CallrFuture}
\title{Prints a callr future}
\usage{
\method{print}{CallrFuture}(x, ...)
}
\arguments{
\item{x}{A CallrFuture object}

\item{\ldots}{Not used.}
}
\description{
Prints a callr future
}
\keyword{internal}
