% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessxFuture-class.R
\name{print.ProcessxFuture}
\alias{print.ProcessxFuture}
\title{Prints a processx future}
\usage{
\method{print}{ProcessxFuture}(x, ...)
}
\arguments{
\item{x}{An ProcessxFuture object}

\item{\ldots}{Not used.}
}
\description{
Prints a processx future
}
\keyword{internal}
