% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallrFuture-class.R
\name{status.CallrFuture}
\alias{status.CallrFuture}
\alias{status}
\alias{finished}
\alias{value}
\title{Status of callr future}
\usage{
\method{status}{CallrFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
A character vector or a logical scalar.
}
\description{
Status of callr future
}
\keyword{internal}
