% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessxFuture-class.R
\name{status.ProcessxFuture}
\alias{status.ProcessxFuture}
\alias{status}
\alias{finished}
\alias{value}
\alias{loggedError}
\alias{loggedOutput}
\title{Status of processx future}
\usage{
\method{status}{ProcessxFuture}(future, ...)
}
\arguments{
\item{future}{The future.}

\item{\ldots}{Not used.}
}
\value{
A character vector or a logical scalar.
}
\description{
Status of processx future
}
\keyword{internal}
