% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallrFutureBackend-class.R
\name{callr}
\alias{callr}
\title{callr futures}
\usage{
callr(..., workers = availableCores(), supervise = FALSE)
}
\arguments{
\item{workers}{The number of processes to be available for concurrent
callr futures.}

\item{supervise}{(optional) Argument passed to \code{\link[callr:r_bg]{callr::r_bg()}}.}

\item{\ldots}{Additional arguments passed to \code{Future()}.}
}
\value{
An object of class \code{CallrFuture}.
}
\description{
A callr future is an asynchronous multiprocess
future that will be evaluated in a background R session.
}
\details{
callr futures rely on the \pkg{callr} package, which is supported
on all operating systems.
}
