% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{future.callr}
\alias{future.callr}
\alias{future.callr-package}
\title{future.callr: A Future for callr}
\description{
The \pkg{future.callr} package implements the Future API
on top of \pkg{callr}.
The Future API is defined by the \pkg{future} package.
}
\details{
To use callr futures, load \pkg{future.callr}, and
select the type of future you wish to use, e.g. \code{plan(callr)}.
}
\examples{
\donttest{
plan(callr)
demo("mandelbrot", package = "future", ask = FALSE)
}

}
